refactor: unix sock disabling proper

+++ darkapis/darkrpc/server.go
@@ -9,6 +9,7 @@ import (
 	"os"
 
 	"github.com/darklab8/fl-darkstat/configs/cfg"
+	"github.com/darklab8/fl-darkstat/darkcore/settings"
 	"github.com/darklab8/fl-darkstat/darkstat/appdata"
 	"github.com/darklab8/fl-darkstat/darkstat/settings/logus"
 )
@@ -48,18 +49,21 @@ var (
 
 func (r *RpcServer) Serve(app_data *appdata.AppData) {
 	rpcServer := rpc.NewServer()
-	rpc_server := NewRpc(app_data)
-	err := rpcServer.Register(rpc_server)
+	rpc_server := NewRpc(app_data)        // NOTE: RPC server is created
+	err := rpcServer.Register(rpc_server) // NOTE: RPC server is registered
 	Log.CheckError(err, "failed to register rpc server")
 
-	rpc.HandleHTTP()                                                   // if serving over http
-	tcp_listener, err := net.Listen("tcp", fmt.Sprintf(":%d", r.port)) // if serving over http
+	rpcServer.HandleHTTP(rpc.DefaultRPCPath, rpc.DefaultDebugPath) // NOTE: Handle path
+	tcp_address := fmt.Sprintf(":%d", r.port)
+	fmt.Println("starting rpc server at ", tcp_address)
+	tcp_listener, err := net.Listen("tcp", tcp_address) // NOTE: listen stuff
 	if err != nil {
 		log.Fatal("listen error:", err)
 	}
 
+	settings.Env.EnableUnixSockets = true
 	var sock_listener net.Listener
-	if cfg.IsLinux {
+	if cfg.IsLinux && settings.Env.EnableUnixSockets {
 		_ = os.Remove(r.sock_address)
 		_ = os.Mkdir("/tmp/darkstat", 0777)
 		sock_listener, err = net.Listen("unix", r.sock_address) // if serving over Unix
@@ -74,7 +78,8 @@ func (r *RpcServer) Serve(app_data *appdata.AppData) {
 	}
 
 	go func() {
-		err := http.Serve(tcp_listener, nil) // if serving over Http
+		fmt.Println("http rpc server is launching")
+		err := http.Serve(tcp_listener, nil) // NOTE: Server
 		if err != nil {
 			log.Fatal("http error:", err)
 		}
